4-project/requirements.txt
pandas
seaborn
sklearn
google-cloud-storage
gcsfs
pydantic
fastapi
uvicorn
transformers